+++ CHANGELOG.md
@@ -85,6 +85,9 @@
 * Fixed an issue where Breadcrumbs were reported in the wrong order on Windows and in the Unity Editor
   [#322](https://github.com/bugsnag/bugsnag-unity/pull/322)
 
+* Fixed an issue where a "Bugsnag loaded" breadcrumb was not added on Windows, Linux, WebGL, and editor builds
+  [#327](https://github.com/bugsnag/bugsnag-unity/pull/327)
+
 * Fixed an issue where the fallback was not reporting the correct app.type
   [#325](https://github.com/bugsnag/bugsnag-unity/pull/325)